move genesisdoc from config to types
