+++ node/full.go
@@ -51,7 +51,7 @@ var _ Node = &FullNode{}
 type FullNode struct {
 	service.BaseService
 
-	genesis *config.GenesisDoc
+	genesis *types.GenesisDoc
 	// cache of chunked genesis data.
 	genChunks []string
 
@@ -79,7 +79,7 @@ func newFullNode(
 	nodeConfig config.NodeConfig,
 	p2pKey crypto.PrivKey,
 	signingKey crypto.PrivKey,
-	genesis *config.GenesisDoc,
+	genesis *types.GenesisDoc,
 	metricsProvider MetricsProvider,
 	logger log.Logger,
 ) (fn *FullNode, err error) {
@@ -171,7 +171,7 @@ func initDALC(nodeConfig config.NodeConfig, logger log.Logger) (*da.DAClient, er
 		namespace, submitOpts, logger.With("module", "da_client")), nil
 }
 
-func initHeaderSyncService(mainKV ds.TxnDatastore, nodeConfig config.NodeConfig, genesis *config.GenesisDoc, p2pClient *p2p.Client, logger log.Logger) (*block.HeaderSyncService, error) {
+func initHeaderSyncService(mainKV ds.TxnDatastore, nodeConfig config.NodeConfig, genesis *types.GenesisDoc, p2pClient *p2p.Client, logger log.Logger) (*block.HeaderSyncService, error) {
 	headerSyncService, err := block.NewHeaderSyncService(mainKV, nodeConfig, genesis, p2pClient, logger.With("module", "HeaderSyncService"))
 	if err != nil {
 		return nil, fmt.Errorf("error while initializing HeaderSyncService: %w", err)
@@ -179,7 +179,7 @@ func initHeaderSyncService(mainKV ds.TxnDatastore, nodeConfig config.NodeConfig,
 	return headerSyncService, nil
 }
 
-func initDataSyncService(mainKV ds.TxnDatastore, nodeConfig config.NodeConfig, genesis *config.GenesisDoc, p2pClient *p2p.Client, logger log.Logger) (*block.DataSyncService, error) {
+func initDataSyncService(mainKV ds.TxnDatastore, nodeConfig config.NodeConfig, genesis *types.GenesisDoc, p2pClient *p2p.Client, logger log.Logger) (*block.DataSyncService, error) {
 	dataSyncService, err := block.NewDataSyncService(mainKV, nodeConfig, genesis, p2pClient, logger.With("module", "DataSyncService"))
 	if err != nil {
 		return nil, fmt.Errorf("error while initializing DataSyncService: %w", err)
@@ -187,7 +187,7 @@ func initDataSyncService(mainKV ds.TxnDatastore, nodeConfig config.NodeConfig, g
 	return dataSyncService, nil
 }
 
-func initBlockManager(signingKey crypto.PrivKey, nodeConfig config.NodeConfig, genesis *config.GenesisDoc, store store.Store, seqClient *seqGRPC.Client, dalc *da.DAClient, logger log.Logger, headerSyncService *block.HeaderSyncService, dataSyncService *block.DataSyncService, seqMetrics *block.Metrics) (*block.Manager, error) {
+func initBlockManager(signingKey crypto.PrivKey, nodeConfig config.NodeConfig, genesis *types.GenesisDoc, store store.Store, seqClient *seqGRPC.Client, dalc *da.DAClient, logger log.Logger, headerSyncService *block.HeaderSyncService, dataSyncService *block.DataSyncService, seqMetrics *block.Metrics) (*block.Manager, error) {
 	exec, err := initExecutor(nodeConfig)
 	if err != nil {
 		return nil, fmt.Errorf("error while initializing executor: %w", err)
@@ -346,7 +346,7 @@ func (n *FullNode) OnStart(ctx context.Context) error {
 }
 
 // GetGenesis returns the genesis document
-func (n *FullNode) GetGenesis() *config.GenesisDoc {
+func (n *FullNode) GetGenesis() *types.GenesisDoc {
 	return n.genesis
 }
