+++ types/utils.go
@@ -15,7 +15,6 @@ import (
 	"github.com/cometbft/cometbft/p2p"
 	cmtypes "github.com/cometbft/cometbft/types"
 	"github.com/libp2p/go-libp2p/core/crypto"
-	"github.com/rollkit/rollkit/config"
 )
 
 // DefaultSigningKeyType is the key type used by the sequencer signing key
@@ -319,7 +318,7 @@ func GetValidatorSetFromGenesis(g *cmtypes.GenesisDoc) cmtypes.ValidatorSet {
 }
 
 // GetGenesisWithPrivkey returns a genesis doc with a single validator and a signing key
-func GetGenesisWithPrivkey(signingKeyType string, chainID string) (*config.GenesisDoc, cmcrypto.PrivKey) {
+func GetGenesisWithPrivkey(signingKeyType string, chainID string) (*GenesisDoc, cmcrypto.PrivKey) {
 	var genesisValidatorKey cmcrypto.PrivKey
 	switch signingKeyType {
 	case "secp256k1":
@@ -329,9 +328,9 @@ func GetGenesisWithPrivkey(signingKeyType string, chainID string) (*config.Genes
 	}
 	pubKey := genesisValidatorKey.PubKey()
 
-	genDoc := config.NewGenesisDoc(chainID, 1, time.Now(), pubKey.Bytes())
+	genesis := NewGenesisDoc(chainID, 1, time.Now(), pubKey.Bytes())
 
-	return genDoc, genesisValidatorKey
+	return genesis, genesisValidatorKey
 }
 
 // PrivKeyToSigningKey converts a privKey to a signing key
